+++ OUDS/Core/Tokens/ComponentTokens/Sources/Values/ControlItemComponentTokens.swift
@@ -69,7 +69,7 @@ public protocol ControlItemComponentTokens {
     var spacePaddingBlockTopHelperText: SpaceSemanticToken { get }
     var spacePaddingBlockBottomSlot: SpaceSemanticToken { get }
     var spacePaddingBlockDensityCompact: SpaceSemanticToken { get }
-    var spacePaddingBlockDensityCompactTopAlignmentTopText_container: SpaceSemanticToken { get }
+    var spacePaddingBlockDensityCompactTopAlignmentTopTextContainer: SpaceSemanticToken { get }
     var spacePaddingBlockDensityCompactTopAlignmentTopCounterweight: SpaceSemanticToken { get }
     var spacePaddingBlockDensityCompactBottomExpandContainer: SpaceSemanticToken { get }
     var spacePaddingBlockDensityDefault: SpaceSemanticToken { get }
@@ -121,6 +121,9 @@ public protocol ControlItemComponentTokens {
     var sizeErrorIcon: SizeSemanticToken { get }
     @available(*, deprecated, message: "Will be removed with next component update")
     var sizeLoader: SizeSemanticToken { get }
+
+    @available(*, deprecated, message: "Use spacePaddingBlockDensityCompactTopAlignmentTopTextContainer instead, will be removed soon")
+    var spacePaddingBlockDensityCompactTopAlignmentTopText_container: SpaceSemanticToken { get }
 }
 
 // swiftlint:enable missing_docs