+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios/compare/2.0.0..develop)
+
+### Changed
+
+- `control item` component token renamed `spacePaddingBlockDensityCompactTopAlignmentTopTextContainer`
+
+### Deprecated
+
+- `control item` component token `spacePaddingBlockDensityCompactTopAlignmentTopText_container`
+
 ## [2.0.0](https://github.com/Orange-OpenSource/ouds-ios/compare/1.4.0...2.0.0) - 2026-05-07
 
 ### Added